# Remove the goroutines from the shortcode lexer